More reorganization
 * Visual reorganization
* Had to move ServerAccordions content into ActiveScriptsPage file. Removed ServerAccordions component.
